Refactored AttackPanelController
